needlectl/requirements.txt
typer
tqdm
requests
pyyaml
pyinstaller